+++ CHANGES.rst
@@ -4,8 +4,9 @@ Changelog
 0.1.1 (????-??-??)
 ------------------
 
-- the readers `from-video-file`, `from-webcam`, `from-youtube`, `from-youtube-live` now use the more efficient
+- the readers `from-video-file`, `from-webcam`, `from-youtube`, `from-youtube-live` now can use a more efficient
   `grab/retrieve` combination instead of `read` to obtain frames, to avoid unnecessary decoding of skipped frames
+  (`--fast`)
 
 
 0.1.0 (2025-10-31)